added comments

+++ MAD-1/Project/application/controllers.py
@@ -1,23 +1,15 @@
-# from ast import Return
-# from calendar import month
-# from crypt import methods
-# from multiprocessing import Value
-# from turtle import color
+import decimal
 from flask import Flask, redirect, request,url_for,render_template,flash
 from flask import current_app as app
-from platformdirs import user_data_dir
 from application.models import *
 from application.api import *
-from application.details import *
-# import urllib.request, json
-# import datetime
 from datetime import datetime
 import requests
 from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy import desc
-# from datetime import datetime
 import matplotlib.pyplot as plt
 from werkzeug.security import generate_password_hash, check_password_hash
+from flask_login import login_required, current_user,login_user,logout_user
 
 
 
@@ -31,7 +23,6 @@ def signup():
 
     if request.method == "POST":
     
-        # created_date= datetime.now()
         user_dict = {
             "user_name" : request.form.get("uname"),
             "user_email" : request.form.get("uemail"),
@@ -47,31 +38,26 @@ def signup():
         #To check password and confirm passwords match and alert user
         if user_dict['user_pwd'] != user_dict['user_cnfmpwd']:
             flash('Passwords doesn\'t Match', category='error')
-            return redirect(url_for('signup'))
-        
         #To check email is valid or not
-        if '@' not in user_dict['user_email']:
+        elif '@' not in user_dict['user_email']:
             flash('Email is not valid', category='error')
-            return redirect(url_for('signup'))
-
-        if user_dict['user_name'].isnumeric():
+        #To check if user name is numeric
+        elif user_dict['user_name'].isnumeric():
             flash('User name should not be numeric', category='error')
-            return redirect(url_for('signup'))
-        
         #To check if already user name exists
-        if user_name:
+        elif user_name:
             flash('Username already exists')
-            return redirect(url_for('signup'))
         #To check if already user email exists
-        if user_email:
+        elif user_email:
             flash('Email address already exists')
-            return redirect(url_for('signup'))
-        
-        #To hash the password and save it in DB
-        user_dict['user_pwd'] = generate_password_hash(user_dict['user_pwd'], method='sha256')
-        #To create new user
-        response_res = requests.post("http://127.0.0.1:5000/v1/api/create", data = user_dict)
-        return render_template("login.html")
+        else:
+            #To hash the password and save it in DB
+            user_dict['user_pwd'] = generate_password_hash(user_dict['user_pwd'], method='sha256')
+            #To create new user
+            response_res = requests.post("http://127.0.0.1:5000/v1/api/create", data = user_dict)
+            flash('User added successfully')
+            return render_template("login.html")
+        return redirect(url_for('signup'))
     return render_template("create.html")
 
 
@@ -91,14 +77,17 @@ def login():
         if not user or not check_password_hash(user.user_pwd, user_pwd):
             flash('Please check your login details and try again.')
             return redirect(url_for('login')) 
+        # if the above check passes, then we know the user has the right credentials
+        login_user(user)
         return redirect(url_for("dashboard",user_id=user.user_id))
         
     return render_template("login.html")
 
 
 @app.route("/logout/<int:user_id>", methods=["GET", "POST"])
+@login_required
 def logout(user_id):
-
+    logout_user()
     response = requests.get(f"http://127.0.0.1:5000/v1/api/logout/{user_id}")
     print("Loggedout")
     return redirect(url_for("home"))
@@ -106,6 +95,7 @@ def logout(user_id):
 
 
 @app.route("/dashboard/<int:user_id>", methods=['GET', 'POST'])
+@login_required
 def dashboard(user_id):
     
         response=requests.get(f"http://127.0.0.1:5000/v1/api/dashboard/{user_id}")
@@ -121,18 +111,18 @@ def dashboard(user_id):
             
             #Dates on Dashboard
             dates=[]
-            # if noof_user_trackers > 1:
+           
             for tracker in user_tracker_details:
                 # print(tracker)
                 islog=Logs.query.filter_by(tracker_id=tracker['tracker_id']).first()
-                print(islog)
+                # print(islog)
                 #To check if atleast one log is present for a tracker
                 if islog is not None:
-                    print("inside is log")
+                    # print("inside is log")
                     last_tracked = Logs.query.filter_by(tracker_id=tracker['tracker_id']).order_by(desc('modified_date')).first()
                     # To check if the log has any modified date, if not it will show the last created date of the log
                     if last_tracked.modified_date is None:
-                        print(last_tracked.created_date)
+                        # print(last_tracked.created_date)
                         last_created_log =  Logs.query.filter_by(tracker_id=tracker['tracker_id']).order_by(desc('created_date')).first()
                         dates.append(last_created_log.created_date)
                     else:
@@ -144,11 +134,13 @@ def dashboard(user_id):
                     # print(type(last_tracked))
                     dates.append(last_tracked)
                 # print(dates)
-            return render_template("dashboard.html",user_data=user_data,user_tracker_details=user_tracker_details,last_tracked=dates)
+            print(current_user.user_name)
+            return render_template("dashboard.html",user_data=user_data,user_tracker_details=user_tracker_details,last_tracked=dates,name=current_user.user_name)
             
             
 
 @app.route(("/addtracker/<int:user_id>"), methods=['GET', 'POST'])
+@login_required
 def addtracker(user_id):
 
     if request.method == "POST":
@@ -160,13 +152,23 @@ def addtracker(user_id):
             "chart_type" : request.form.get("chart_type"),
             "settings" : request.form.get("settings")
         }
-        # print(add_dict)
-        response = requests.post(f"http://127.0.0.1:5000/v1/api/addtracker/{user_id}", data=add_dict)
-        print(response)
-        return redirect(url_for("dashboard",user_id=user_id))
+        print(add_dict)
+        
+        # To check if user has the same tracker name already
+        user_tracker_names=Tracker.query.filter_by(name=add_dict['name'],user_id=user_id).first()
+        if user_tracker_names:
+            flash('Tracker Already Exist.Try another Name')
+        #To check if tracker name is numeric
+        elif add_dict['name'].isnumeric():
+            flash('Tracker Name should not be numeric')
+        else:  
+            response = requests.post(f"http://127.0.0.1:5000/v1/api/addtracker/{user_id}", data=add_dict)
+            return redirect(url_for("dashboard",user_id=user_id))
+        return redirect(url_for('addtracker',user_id=user_id))
     return render_template('addtracker.html',user_id=user_id)
 
 @app.route(("/updatetracker/<int:user_id>/<int:tracker_id>"), methods=['GET', 'POST'])
+@login_required
 def updatetracker(user_id,tracker_id):
     
     if request.method == "POST":
@@ -177,32 +179,44 @@ def updatetracker(user_id,tracker_id):
             "chart_type" : request.form.get("chart_type"),
             "settings" : request.form.get("settings")
         }
+
+        #To check if tracker name is numeric
+        if add_dict['name'].isnumeric():
+            flash('Tracker Name should not be numeric')
+            return redirect(url_for('updatetracker',user_id=user_id,tracker_id=tracker_id))
+
         response = requests.put(f"http://127.0.0.1:5000/v1/api/updatetracker/{user_id}/{tracker_id}", data=add_dict)
-        print(response)
-        return redirect(url_for("dashboard",user_id=user_id))
-    tracker_data = Tracker.query.filter_by(tracker_id=tracker_id).first()
-    print(tracker_data.settings)
-    
+        
+        return redirect(url_for("dashboard",user_id=user_id,name=current_user.user_name))
+
+    # GET method code
+    tracker_data=Tracker.query.filter_by(tracker_id=tracker_id).first()
     response=requests.get(f"http://127.0.0.1:5000/v1/api/dashboard/{user_id}")
     user_data=json.loads(response.text)
-    return render_template('updatetracker.html',tracker_data=tracker_data,user_data=user_data)
+
+    return render_template('updatetracker.html',tracker_data=tracker_data,user_data=user_data,name=current_user.user_name)
 
 
 
 @app.route(("/deletetracker/<int:user_id>/<int:tracker_id>"), methods=['GET', 'POST'])
+@login_required
 def deletetracker(user_id,tracker_id):
+    #To delete the tracker data
     response = requests.delete(f"http://127.0.0.1:5000/v1/api/deletetracker/{user_id}/{tracker_id}")
-    print(response)
     response=requests.get(f"http://127.0.0.1:5000/v1/api/log_data/{user_id}/{tracker_id}")
-    log_data=json.loads(response.text)
+    
+    if response.status_code != 404:
+        log_data=json.loads(response.text)
+        print(log_data)
+        #To delete the logs for the tracker
+        for log in log_data:
+            response = requests.delete(f"http://127.0.0.1:5000/v1/api/deletelog/{user_id}/{tracker_id}/{log['log_id']}")
 
-    for log in log_data:
-        response = requests.delete(f"http://127.0.0.1:5000/v1/api/deletelog/{user_id}/{tracker_id}/{log['log_id']}")
-        print(response)
-    return redirect(url_for("dashboard",user_id=user_id))
+    return redirect(url_for("dashboard",user_id=user_id,name=current_user.user_name))
 
 
 @app.route(("/logevent/<int:user_id>/<int:tracker_id>"), methods=['GET', 'POST'])
+@login_required
 def logevent(user_id,tracker_id):
 
     if request.method == "POST":
@@ -213,35 +227,57 @@ def logevent(user_id,tracker_id):
                 "notes" : request.form.get("notes"),
                 "selected_choice" : request.form.get("multiple_type")
             }
-        # print(add_log_dict)
+        print(add_log_dict)
+        def isfloat(num):
+            try:
+                float(num)
+                return True
+            except ValueError:
+                return False
+        if not isfloat(add_log_dict['value']):
+            flash('Value should be numeric')
+            return redirect(url_for('logevent',user_id=user_id,tracker_id=tracker_id))
+        
         add_log_dict["log_time"]=datetime.strptime(add_log_dict["log_time"], '%Y-%m-%dT%H:%M')
         response= requests.post(f"http://127.0.0.1:5000/v1/api/loganewevent/{user_id}/{tracker_id}",data=add_log_dict)
-        return redirect(url_for("viewtracker",user_id=user_id,tracker_id=tracker_id))
+        return redirect(url_for("viewtracker",user_id=user_id,tracker_id=tracker_id,name=current_user.user_name))
+    
+    #To display Multiple choice values on screen
     tracker_data = Tracker.query.filter_by(tracker_id=tracker_id).first()
-    print(tracker_data.type)
+    # print(tracker_data.type)
     multiple_data=[]
     if tracker_data.type == "MultipleChoice":
-        
         for d in tracker_data.settings.split(","):
             multiple_data.append(d)
-    return render_template("logevent.html",user_id=user_id,tracker_id=tracker_id,multiple_data=multiple_data,tracker_data=tracker_data)
+
+    return render_template("logevent.html",user_id=user_id,tracker_id=tracker_id,multiple_data=multiple_data,tracker_data=tracker_data,name=current_user.user_name)
 
 @app.route(("/updatelog/<int:user_id>/<int:tracker_id>/<int:log_id>"), methods=['GET', 'POST'])
+@login_required
 def updatelog(user_id,tracker_id,log_id):
 
-    update_dict = {
+    if request.method == "POST":
+
+        update_dict = {
         "log_time" : request.form.get("ltime"),
         "value" : request.form.get("lvalue"),
         "notes" : request.form.get("notes"),
         "selected_choice" : request.form.get("multiple_type")
-    }
-    # print(update_dict)
-    
-    if request.method == "POST":
+        }
+        # print(update_dict)
+        def isfloat(num):
+            try:
+                float(num)
+                return True
+            except ValueError:
+                return False
+        if not isfloat(update_dict['value']):
+                flash('Value should be numeric')
+                return redirect(url_for('updatelog',user_id=user_id,tracker_id=tracker_id,log_id=log_id))
         update_dict['log_time'] = datetime.strptime(update_dict["log_time"], '%Y-%m-%dT%H:%M')
         response = requests.put(f"http://127.0.0.1:5000/v1/api/update_log_data/{user_id}/{tracker_id}/{log_id}",data=update_dict)
-        print(response.text)
-        return redirect(url_for("viewtracker",user_id=user_id,tracker_id=tracker_id))
+        # print(response.text)
+        return redirect(url_for("viewtracker",user_id=user_id,tracker_id=tracker_id,name=current_user.user_name))
         
     response=requests.get(f"http://127.0.0.1:5000/v1/api/update_log_data/{user_id}/{tracker_id}/{log_id}")
     update_log_data=json.loads(response.text)
@@ -250,45 +286,48 @@ def updatelog(user_id,tracker_id,log_id):
     multiple_data=[]
     for d in tracker_data.settings.split(","):
         multiple_data.append(d)
-    print(update_log_data)
-    return render_template("updatelog.html",log_data=update_log_data,user_id=user_id,tracker_id=tracker_id,log_id=log_id,tracker_data=tracker_data,multiple_data=multiple_data)
+    # print(update_log_data)
+    return render_template("updatelog.html",log_data=update_log_data,user_id=user_id,tracker_id=tracker_id,log_id=log_id,tracker_data=tracker_data,multiple_data=multiple_data,name=current_user.user_name)
 
 
 @app.route(("/deletelog/<int:user_id>/<int:tracker_id>/<int:log_id>"), methods=['GET', 'POST'])
+@login_required
 def deletelog(user_id,tracker_id,log_id):
     response = requests.delete(f"http://127.0.0.1:5000/v1/api/deletelog/{user_id}/{tracker_id}/{log_id}")
-    print(response)
-    return redirect(url_for("viewtracker",user_id=user_id,tracker_id=tracker_id))
+    # print(response)
+    return redirect(url_for("viewtracker",user_id=user_id,tracker_id=tracker_id,name=current_user.user_name))
 
 @app.route(("/userlogs/<int:user_id>/<int:tracker_id>"), methods=['GET', 'POST'])
+@login_required
 def viewtracker(user_id,tracker_id):
 
     response=requests.get(f"http://127.0.0.1:5000/v1/api/log_data/{user_id}/{tracker_id}")
     print(response.status_code)
+    
     tracker_data = Tracker.query.filter_by(tracker_id=tracker_id).first()
     islog=Logs.query.filter_by(tracker_id=tracker_id).count()
     if response.status_code==404:
         
-        return render_template("tracker.html",user_id=user_id,tracker_id=tracker_id,tracker_data=tracker_data,islog=islog)
+        return render_template("tracker.html",user_id=user_id,tracker_id=tracker_id,tracker_data=tracker_data,islog=islog,name=current_user.user_name)
     else:
         log_data=json.loads(response.text)
         last_tracked = Logs.query.order_by(desc('modified_date')).first()
         # print(last_tracked.modified_date)
         
         tracker_data = Tracker.query.filter_by(tracker_id=tracker_id).first()
-        print(tracker_data.type)
+        # print(tracker_data.type)
         x=[]
         y=[]
         if tracker_data.type == 'Timestamp':
             for log in log_data:
                 x.append(log["log_time"])
-                y.append(int(log["value"]))   
+                y.append(float(log["value"]))   
         
         dict_data={}
         if tracker_data.type == 'Numeric':
             for log in log_data:
                 x.append(log["notes"])
-                y.append(int(log["value"])) 
+                y.append(float(log["value"])) 
             
             for data in list(zip(x, y)):
                 if data[0] in dict_data.keys():
@@ -303,7 +342,7 @@ def viewtracker(user_id,tracker_id):
         if tracker_data.type == 'MultipleChoice':
             for log in log_data:
                 x.append(log["selected_choice"])
-                y.append(int(log["value"]))
+                y.append(float(log["value"]))
             # print(list(zip(x,y)))
             for data in list(zip(x,y)):
                 if data[0] in dict_data.keys():
@@ -332,20 +371,20 @@ def viewtracker(user_id,tracker_id):
         plt.savefig('static/img/logplot.png',dpi=70,bbox_inches="tight")
         plt.clf()  
 
-    return render_template("tracker.html",user_id=user_id,log_data=log_data,tracker_id=tracker_id,last_tracked=last_tracked.modified_date,tracker_data=tracker_data,islog=islog)
+    return render_template("tracker.html",user_id=user_id,log_data=log_data,tracker_id=tracker_id,last_tracked=last_tracked.modified_date,tracker_data=tracker_data,islog=islog,name=current_user.user_name)
 
 
 
 
 
 @app.route("/profile/<int:user_id>", methods=['GET', 'POST'])
+@login_required
 def user_profile(user_id):
-
+    user_profile = requests.get(f"http://127.0.0.1:5000/v1/api/user/{user_id}")
+    user_data=json.loads(user_profile.text)
+    # print(user_data)
     if request.method == "GET":
-
-        user_profile = requests.get(f"http://127.0.0.1:5000/v1/api/user/{user_id}")
-        user_data=json.loads(user_profile.text)
-        return render_template("myprofile.html",user_data=user_data)
+        return render_template("myprofile.html",user_data=user_data,name=current_user.user_name)
 
     if request.method == "POST":
 
@@ -356,13 +395,21 @@ def user_profile(user_id):
             "sec_question" : request.form.get("questions"),
             "sec_answer" : request.form.get("ans")
         }
-        user_json_object = json.dumps(user_dict, indent = 4)
-    
-        print(user_json_object)
-        
-        user_update = requests.put(f"http://127.0.0.1:5000/v1/api/user/{user_id}", data = user_dict)
-       
-        return render_template("user_profile_update.html", user_id=user_id)
+        # print(user_dict)
+        if user_dict['user_pwd'] != user_dict['user_cnfmpwd']:
+            flash("New Password and Confirm Password doesn't match.")
+        elif not check_password_hash(user_data['user_pwd'], user_dict['user_pwd1']):
+            flash("Old Password is incorrect. Please try again")
+        else:
+            user_dict['user_pwd'] = generate_password_hash(user_dict['user_pwd'], method='sha256')
+            user_update = requests.put(f"http://127.0.0.1:5000/v1/api/user/{user_id}", data = user_dict)
+            flash("Profile updated successfully")
+            return redirect(url_for('dashboard',user_id=user_id))
+        return redirect(url_for('user_profile',user_id=user_id))
+
+@app.route('/aboutus')
+def aboutus():
+    return render_template('aboutus.html')
